First version of mls-lite prototype
The code here exposes a minimal C API for

* Generating key packages
* Creating and joining groups
* Adding and removing group members

The API is built around a very simple idea: we wrap the main types
from mls-rs in newtypes and expose just the methods we need to
implement the critical user journeys identified by Anh and Giles.
Mirroring the mls-rs API makes this easy: we have the right types at
hand when needed and most methods will closely resemble the methods
from mls-rs, but perhaps with some arguments removed and replaced with
hard-coded defaults.

The storage is in-memory right now: the idea is to pass in callback
function(s) to `Client::new` which will then be used to implement
`GroupStateStorage`.

Note also that the code here is using safer-ffi like mls-rs currently
does. This means that you get a very low-level C-compatible API layer.
The tests show that it’s possible to use this layer, which means we
can also call the functions from Python, Kotlin, Swift, etc…

However, it would be nicer to expose a more high-level API via UniFFI.
I’ve started playing with this, but it’s very early days and I’m still
trying to figure out how UniFFI works.

I added the auto-generated `mls-lite.h` and `mls-lite.rs` files at the
root: that way `cargo test` should work for you out of the box.
Long-term, we will not use these files assuming that we switch to
UniFFI instead.