Add methods to `LiteClient` and `LiteGroup`
This highlighted an interesting limitation of UniFFI APIs: because it
makes the APIs memory safe even when the foreign language holds a
pointer to a Rust value, methods cannot take `&mut self` (we cannot
guarantee exclusive access).

The result of this is heavy use of interior mutability with lots of
`Arc<Mutex<T>>` wrapping. It clutters the mls-lite API a lot, but it
seems to be tolerable right now.