Make methods async
Supporting both sync and async code using maybe-async seems infeasible
right now. Exporting the methods using UniFFI requires:

    #[cfg_attr(feature = "uniffi", uniffi::export)]
    impl LiteClient {

However, maybe-async needs per-method annotations. It looks like these
annotations are processed _after_ UniFFI sees the code (the order of
proc-macro expansion is not yet defined, but it would make sense to
expand them from the outside-in).

For now, we can let the code be async and figure out how to deal with
this later.